+++ spec/Pim/Bundle/DataGridBundle/Datasource/DatasourceSpec.php
@@ -8,12 +8,13 @@ use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Doctrine\ORM\ProductMassActionRepository;
 use Pim\Bundle\CatalogBundle\Doctrine\ORM\ProductRepository;
 use Pim\Bundle\DataGridBundle\Datasource\ResultRecord\HydratorInterface;
+use Pim\Bundle\CatalogBundle\Doctrine\Query\ProductQueryFactoryInterface;
 
 class DatasourceSpec extends ObjectBehavior
 {
-    function let(ObjectManager $manager, HydratorInterface $hydrator, ProductMassActionRepository $massActionRepo)
+    function let(ObjectManager $manager, HydratorInterface $hydrator, ProductMassActionRepository $massActionRepo, ProductQueryFactoryInterface $factory)
     {
-        $this->beConstructedWith($manager, $hydrator, $massActionRepo);
+        $this->beConstructedWith($manager, $hydrator, $massActionRepo, $factory);
     }
 
     function it_is_a_datasource()
@@ -31,7 +32,7 @@ class DatasourceSpec extends ObjectBehavior
             'entity'            => 'Product'
         ];
         $manager->getRepository('Product')->willReturn($repository);
-        $repository->createDatagridQueryBuilder()->shouldBeCalled();
+        $repository->createDatagridQueryBuilder([])->shouldBeCalled();
         $grid->setDatasource($this)->shouldBeCalled();
         $this->process($grid, $config);
     }
@@ -61,7 +62,7 @@ class DatasourceSpec extends ObjectBehavior
             'entity' => 'Product'
         ];
         $manager->getRepository('Product')->willReturn($repository);
-        $repository->createQueryBuilder('o')->shouldBeCalled();
+        $repository->createQueryBuilder([])->shouldBeCalled();
         $grid->setDatasource($this)->shouldBeCalled();
         $this->process($grid, $config);
     }