+++ spec/Pim/Bundle/FilterBundle/Filter/Product/CompletenessFilterSpec.php
@@ -4,8 +4,6 @@ namespace spec\Pim\Bundle\FilterBundle\Filter\Product;
 
 use Doctrine\ORM\QueryBuilder;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Doctrine\Query\ProductQueryBuilderInterface;
-use Pim\Bundle\CatalogBundle\Repository\ProductRepositoryInterface;
 use Pim\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
 use Pim\Bundle\FilterBundle\Filter\ProductFilterUtility;
 use Symfony\Component\Form\FormFactoryInterface;
@@ -24,30 +22,18 @@ class CompletenessFilterSpec extends ObjectBehavior
 
     function it_applies_a_filter_on_complete_products(
         FilterDatasourceAdapterInterface $datasource,
-        $utility,
-        ProductRepositoryInterface $repository,
-        ProductQueryBuilderInterface $pqb,
-        QueryBuilder $qb
+        $utility
     ) {
-        $datasource->getQueryBuilder()->willReturn($qb);
-        $utility->getProductRepository()->willReturn($repository);
-        $repository->getProductQueryBuilder($qb)->willReturn($pqb);
-        $pqb->addFilter('completeness', '=', 100)->shouldBeCalled();
+        $utility->applyFilter($datasource, 'completeness', '=', 100)->shouldBeCalled();
 
         $this->apply($datasource, ['type' => null, 'value' => 1]);
     }
 
     function it_applies_a_filter_on_not_complete_products(
         FilterDatasourceAdapterInterface $datasource,
-        $utility,
-        ProductRepositoryInterface $repository,
-        ProductQueryBuilderInterface $pqb,
-        QueryBuilder $qb
+        $utility
     ) {
-        $datasource->getQueryBuilder()->willReturn($qb);
-        $utility->getProductRepository()->willReturn($repository);
-        $repository->getProductQueryBuilder($qb)->willReturn($pqb);
-        $pqb->addFilter('completeness', '<', 100)->shouldBeCalled();
+        $utility->applyFilter($datasource, 'completeness', '<', 100)->shouldBeCalled();
 
         $this->apply($datasource, ['type' => null, 'value' => 2]);
     }