+++ spec/Pim/Bundle/FilterBundle/Filter/Product/EnabledFilterSpec.php
@@ -4,8 +4,6 @@ namespace spec\Pim\Bundle\FilterBundle\Filter\Product;
 
 use Doctrine\ORM\QueryBuilder;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Doctrine\Query\ProductQueryBuilderInterface;
-use Pim\Bundle\CatalogBundle\Repository\ProductRepositoryInterface;
 use Pim\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
 use Pim\Bundle\FilterBundle\Filter\ProductFilterUtility;
 use Symfony\Component\Form\FormFactoryInterface;
@@ -24,15 +22,9 @@ class EnabledFilterSpec extends ObjectBehavior
 
     function it_applies_a_filter_on_enabled_field_value(
         FilterDatasourceAdapterInterface $datasource,
-        $utility,
-        ProductRepositoryInterface $repository,
-        ProductQueryBuilderInterface $pqb,
-        QueryBuilder $qb
+        $utility
     ) {
-        $datasource->getQueryBuilder()->willReturn($qb);
-        $utility->getProductRepository()->willReturn($repository);
-        $repository->getProductQueryBuilder($qb)->willReturn($pqb);
-        $pqb->addFilter('enabled', '=', 1)->shouldBeCalled();
+        $utility->applyFilter($datasource, 'enabled', '=', 1)->shouldBeCalled();
 
         $this->apply($datasource, ['type' => null, 'value' => [0 => 1]]);
     }