+++ src/Pim/Bundle/FilterBundle/Filter/ProductValue/ChoiceFilter.php
@@ -68,8 +68,8 @@ class ChoiceFilter extends AjaxChoiceFilter
         $this->util->applyFilter(
             $ds,
             $this->get(ProductFilterUtility::DATA_NAME_KEY),
-            $data['value'],
-            $operator
+            $operator,
+            $data['value']
         );
 
         return true;